Extract hive compatibility test
The test `testInsertAllSupportedDataTypesWithTrino` from TestHiveStorageFormats
is being executed against several Hive distributions.
In order to cope easier with the limited capabilities of the Hive distribution
against which the test is being executed, it is best to extract it to a separate
class for allowing an easier extension of the test.
